Ensure that AssetCreator is not used before proto.LogicSigVersion=5 (#2784)
This also unifies a lot of field handling so that the code is less
susceptible to copy-paste errors, but it has a ways to go. Hard to
both unify the handling of the different fields and use separate
types for each field index. Lots and lots of interfaces could do. Or
generics in 6 months? For the moment, each field type is a pile of
arrays and maps working together.

It's unclear what we gain from those separate index types, since
we always cast into the type just before using them, as we are
getting the indexes from reading a byte in bytecode.

New tests will catch this mistake if we make it again.