Ensure disassemble/reassemble cycle works in testProg. (#2745)
Disassemble was omitting labels if they were only used as targets of backjumps.

This fixes that, but making a second pass to find labels if any such labels exist in a program. It also changes testProg, which is used extensively to test programs in the logic package, so that every such test is a Disassemble/Reassemble test to ensure the same program is produced.