fix TestAgreementSynchronousFutureUpgrade with pipelining
that test required us to accept votes for a future round, before we have
even seen the proposal payload for the previous round.  this failed
because the future round's Branch was not known to the speculative
ledger yet.  the fix is to use LedgerBranchReader and to explicitly
attach the Branch value to use to the messageEvent.