Bugfix/fix nano s enumeration bug (#2822)
Goal was not able to communicate with the Ledger Nano S device.
The device has two HID occurrences on the OS and after the firmware upgrade, one of them became not responsive. That made goal to the hang.

fixing the interaction with the Ledger devices (S and X) on the Windows platform.