Use math.bits intrinsics to simplify addw/mulw opcodes (#2839)
Removes some code, but more importantly speeds up mulw considerably.
We have always charged 1 unit of cost for mulw, but it was actually
more expensive.  This fixes that.

A nice side-effect is that opPlus/opAddw and opMul/opMulw are
simplified and easier to follow.  Benchmarks confirm nothing is
slower (and mulw is ~5 times faster).

All tests of opAddwImpl and opMulwImpl have been retained by
conducting the same test with actual teal, instead of directly testing
those (now non-existent) functions.