Fix partition recovery tests (#2820)
The expected round number is captured before the node stops. However,
it is likely that the node advances to the next round before it is
stopped. When this happens, the test will fail.

This change gets the most up-to-date round number after the node is
stopped, but before inducePartitionTime timeout is waited.

inducePartitionTime is the wait to make sure the expected behavior is
obtained. The round number is captured before this wait.

However, I could not identify in this PR why TestBasicPartitionRecovery has failed. Could not find anything in the test, and the failure logs have nothing.
I suspect that the failure in the other tests triggered the failure, and fixed the other tests, but cannot be sure.

As for the data race, it is fixed in #2844.

Fixes #2384 and #2545