txnsync: bloom filter refactoring (#2834)
Refactor the bloom filter implementation to have two "implementations" : testableBloomFilter and bloomFilter.

testableBloomFilter is used exclusively when being received from the network; it does not contain any encoded
fields, nor does it include transactionsRange which is used for generating new bloom filters.

bloomFilter on the flip side is used only for outgoing bloom filters. It doesn't support testing, and it contains only the encoded bloom filter ( and not the underlying bloom filter needed for testing, for instance ).

This separation allows us to ensure we don't store bloom filter object beyond the required scope, improving the memory utilization.