Display app address in `goal app info` & show logs are binary in REST API (#2836)
This is a follow up to #2661. It adds the app account address to the goal app info command, and it changes the log item type format in the REST API to "byte" in order to show that it's binary data. As a bonus, the encoder will automatically convert to base64 for us.