Increment TxnCounter with inner transactions
This only changes the blocks, so there is still a lot of work to do so
that external APIs see things properly.  The REST API assumes that it
can determine a creatble ID based on the index of a transaction in a
block and the block's TxnCounter.  That is no longer true, it must
actually figure out how many inner txns it is skipping over.