fix multiple timeouts for the same instant
when we wanted multiple timeouts that happen to be at the same time,
our implemenation had funny behavior: if the timeout already expired,
we got a closed channel, which implicitly signaled as many timeouts as
we wanted.  but if the timeout was in the future, time.After() would
only write a single time.Time value in the channel, so at most one
timeout would fire.