fix EvalForIndexer (#2941)
Summary
In the EvalForIndexer, we don't want to call finalValidation(). This would
skip the calculation of the account totals in the state delta, which is a serious
issue if it were to be used by algod, but it's perfectly fine for the indexer since
it doesn't track any totals and therefore cannot calculate the new totals.

Test Plan
Test against the indexer.