Fix for TestKeysWithoutStateProofKeyCannotRegister and TestAccountStorageWithBlockProofID (#3076)
Fix for TestKeysWithoutStateProofKeyCannotRegister and TestAccountStorageWithBlockProofID

The test TestKeysWithoutStateProofKeyCannotRegister lets the nodes make progress until changes the consensus version.
The test passes when the consensus version changes to Future. However, the test starts from V29, and quickly gets to Future, passing through V30. Sometimes, the node progress is halted at V30, in which case the test fails.

Also, in this change, reusing the client handle instead of regenerating one every time.

TestAccountStorageWithBlockProofID broke after the merge with changes to the Master.
This fixes it.