txnsync: implement missing bridge between txnsync and classic transaction messages (#3102)
Summary
This PR adds the missing bridge between the txnsync and the classic transaction relaying:
when a transaction message arrive and being added to the transaction pool, we need to attempt to
relay these messages right away using the classic transaction messages. That would allow relays to
be compatible with both 2.1 and 3.0 nodes.

Test Plan
e2e test was added.