Add c5d.4xlarge and c5d.18xlarge us-east-2 to hosttemplates.json (#3160)
Add c5d.4xlarge and c5d.18xlarge us-east-2 to hosttemplates.json so that our team can use them moving forward for performance tests.

For performance testing, we've needed to add these to each codebase we're working with to use these machines. This is part of our effort to automate performance testing internally.