interactive mode for e2e testing
Some e2e tests require a python environment for testing.
Unfortunately, setting up that environment adequately similar to the testing environment may not be trivial.
This change introduces an interactive mode for the e2e.sh script which stops at the point of running the tests, and allows the user to run the tests from the same testing environment.