Adding missing msgp bounds (#3196)
The following slices are not bounded for msgp:
- OneTimeSignatureSecretsPersistent Batches, Offsets
- ByteSignature

OneTimeSignatureSecretsPersistent Batches, Offsets 
These are not bounded because the bound is keyDilution, which is a settable parameter. Moreover, the decoding of these structures does not happen for a message coming from an untrusted remote source. 

ByteSignature
The bound set is the larger of the two signature schemes used: Falcon signature size.