ledger refactoring: catchpoint fixes (#3374)
## Summary

Fixes to `processAllResources` and `processAllBaseAccountRecords` related to a case when an account does not have resource records and a pending row needs to be carried over next account or a batch reading iteration.

Use `processAllBaseAccountRecords` in `encodedAccountsBatchIter.Next()`. This allows us to the same code for all scenarios where we read entire base account and resources tables.