agreement: Add validatedAt to proposal type and BlockAcceptedEventDetails message (#3703)
## Summary

This pulls a bit of code from the the [pipelining branch](https://github.com/algorand/go-algorand/tree/feature/pipeline) to add a "validatedAt" duration field to the agreement.proposal type. This is used to provide an extra field in the BlockAcceptedEventDetails message describing how soon into the round each block was validated.

## Test Plan

Existing tests should pass, and a new test might be helpful to show that the value is getting filled in correctly.