fast catchup: retry peers fetching (#3711)
Summary

The existing code was failing the fast catchup on node startup if the peers could not be retrieved from the network package.
Unfortunately, this is almost always the case. This change allow the node to retry fetching the peers list from the network package multiple times, while having a short delay between the iterations.

This issue was reported by a Algorand Forum member.

Test Plan

Test the change manually.