ledger: test OnlineAccountData
* This is a prerequisite test before retiring AccountData.OnlineAccountData()
that is used only in tests agreement and committee tests.
* Historically all these tests depends on genesis data that is basic.AccountData.
* Our GA genesis files have stake, status and voting data and never had resources
basic.AccountData can be reduced to the original stake + voting data state and
a new basic.AccountDataEx can be introduced for using in REST API
* Ledger tests and test network generation tools need to be updated to
accept a new reduced (original) basic.AccountData as genesis and resource records
for assets and apps needed for tests after the ledger initialization.