ledger: fix possible dbRound sync issue in trackers and the registry (#3910)
The trackerRegistry.dbRound value and cached dbRound values stored in trackers
might go out of sync. Although they are updated under the same lock,
produceCommittingTask might use outdated dbRound and give it to trackers
with the updated state.
The fix is simple: to have dbRound usage and produceCommittingTask invocation
under the same lock.