AVM: inner appls can call v4 (#3896)
AllowV4InnerAppls also means that apps can not perform an inner opt-in
to apps with <v4 CSP (since they might then be locked in). And no app
can downgrade, as that is a straightforward rule to prevent an app
from lowering its CSP and locking other apps into it.

The DoubleLedger does a lot of double checking on ledgers produced by
rerunning transactions after creating blocks, deserializing the txns
from the block and validating the results.

testConsensusRange makes it easy for ledger tests to test behavior in
many protocol levels. A particular problem it handles is that test can
be written to run from some consensus version and in every subsequent
version. So if we change behavior in a new version, we'll find out.


Co-authored-by: algochoi <86622919+algochoi@users.noreply.github.com>