Make "0 index" box refs on app calls available in group
Gooing with an idea from Michael to help regularize how to make box
refs available to the entire group if it is an app create, and a "0
index" box ref. At ep creation time, we don't know the appId, so it
can't go into ep.available.boxes.  But when we begin evaluating the
txn, we do know the appID. So add it then. Doing the addition at the
start of EvalContract, rather than in RecordAD allows the creation
call itself to access the box - though that is incredibly unlikely to
be useful, it seemed proper. (see boxtxn_test.go for explanation of
how to take advantage of this.)