Allow "empty" accounts to be Senders if someone else pays
The change affects any Move() operation - if the amount moved is 0,
and the account experiencing the move does not have enough algos for
the rewardsLevel increment to matter, skip it.

This has the haoppy side-effect of preventing some unneccesary db
writes, but the main advantage is clarity of specification. The
previous implementation did not have clear semantics, as the effects
of a transfer dependened on whether the network was new (no rewards
paid yet) or not.