Allow "empty" accounts to be Senders if someone else pays (#4151)
The change affects any Move() operation - if the amount moved is 0,
and the account experiencing the move does not have enough algos for
the rewardsLevel increment to matter, skip it.

This has the happy side-effect of preventing some unnecessary db
writes, but the main advantage is clarity of specification. The
previous implementation did not have clear semantics, as the effects
of a transfer depended on whether the network was new (no rewards
paid yet) or not.