dupn and bury
Also, a lot of cleanup of type tracking with frame_*

And, caught some examples of us not testing opcodes in "nonsense"
tests because the opcode was a substring of another opcode, like & is
a substring of &&. (Caught, because "bury" is a substring of
"frame_bury")

This means there is a change to the "old" compiled constants.  This is
normally not done, but here we added to an old nonsense test, we added
the corresponding bytes to the compiled output.