Consolidate models to a single package
Moves all types.go to a single model package within the v2 generated API code. Also
moves the oapi-codegen config file for each of the generated packages alongside the code
it is generating. Additionally, we've moved the generated routes.go into a public package
instead of the top level generated package.

The result of this is a large refactoring of imports. generated=>models,
generatedV2=>models, and generated=>public in the case of the router.