API: Update oapi-codegen version to v1.12.0 of our fork (#4707)
Updates the version of oapi-codegen which is installed via
the buildtools to our v1.12.0 release. Also changes the
Makefile targets and adds config files which are the preferred
way of configuration instead of the old CLI flags.

The generated types definitions have also been updated using
the oapi-codegen commit which ensures all required Enums are generated
for enums embedded in response types.