Fix Unstable Unit Test; Post-Feature-Merge Cleanup
Addresses a flaky Unit Test by not overloading the component under test
(previously, 5k messages were driven as-fast-as-possible. Now, 10k
messages are driven in batches of 500 with a 100ms delay). The new
behavior is more accurate to a realistic usage pattern, as well.

This also addresses some deletable code and adds comments where
requested from some post-merge requests on pull #4797.