Add addPeerLockless for consistent peersByID checking
Previously in tryConnect, the peerLock is taken to check the peersByID
map for potential deduplication before calling addPeer. Once the check
is done, there is a window where the value in the peersByID map may
change before the call to addPeer takes the lock again.

To address this, addPeerLockless as been added, and addPeer manages the
locks around this function. In tryConnect now, if there is potential
deduplication (if the peer has a verified identity), the lock is held
from the time the peersByID map is checked, through and including adding
the peer. tryConnect will return the lock as soon as possible.

If the peer being added does *not* have a verified identity, the lock is
never taken and the standard addPeer is called, because there is no
deduplication to be done.