Move Identity Tracking into an identityTracker struct
Now, the management of the "peersByID" map is done by a struct which
manages its own locking and can be unit tested apart from wsNetwork
tests. This reverts the need for an addPeerLockless function.

This also incorporates more comments from cce, so now wn.Start is also
back to its original form.