ledger: report catchpoint writing only when it actually started
The original implementation was setting this flag in task producer,
that created a gap between setting it and actual writing catchpoint
in commitSyncer.

Later, #5214 introduced ability to skip tasks if the queue is full.
This caused an issue with catchup service that was reading the flag
with IsWritingCatchpointDataFile() and stopping the catchup.
Because of that ledger was not receiving new blocks and was unable
to schedule a new commit task but had the catchpoint writing flag set
from the previous discarded task.