Simplfy conversion and fix a spurious complaint from static analysis
Static analysis tools noted that:
```
 	for i, sv := range cx.stack {
-		stack[i] = stackValueToTealValue(&sv)
```
was a potential problem because `sv` is always the same variable in a
for loop, so this code was passing the same address on each call.
That was fine, since stackValueToTealValue is single-threaded, and
does not sture the address. It's just a simple conversion.

But, noting that the conversion was more complicated than it needed to
be, and probably allocated more than it needed to, this PR simplifies
it and makes it work on stackValue values instead of pointers.