Make apps and app accounts available while creation is "pending"
Starting in the same consensus version that allows resource sharing,
this PR makes apps and app accounts of "pending" app creates available
while their inners are executing.

This was the intent all along, the new app was only added to
`cx.createdApps` at the conclusion of its execution, rather than at
the start.