Small cleanups for clarity
I'm doing a larger bit of work to enable logicsig cost pooling and I
found myself making these changes along the way. It's easier to review
by extracting them into their own PR.  This does three things.

Introduces and uses:
```
func (msig MultisigSig) Signatures() int
```
for the sevral places we need to examine a MultiSig to count the
number of sub signatures present.

Uses the pre-existing `func (s *Signature) Blank() bool` in many
places.

Renames `EvalContext.runModeFlags` to simply `EvalContext.runMode`
because it takes on a single value, it can't be, for example: `ModeSig |
ModeApp`, it must be one or the other.