add block hash to state proof
Replace the `Seed` value in the state proof light header with `BlockHash`.
It serves a similar purpose in terms of the second-order consideration
of defending against potential future hash-collision attacks (albeit a
little bit weaker), but has the benefit of allowing validating the entire
blockchain, by committing to the block hash itself in the state proof.