fix TestStateProofMessageCommitmentVerification
The test set a configurable consensus protocol, "test-fast-stateproofs",
and propagated the configuration to the child algod processes spawned
by the test, but did not update the supported consensus params in
the test process itself.  As a result, when the test code tried to
compute the light block header, it didn't have the block protocol
(test-fast-stateproofs) in its config.Consensus.