ledger: fix rare race in TestCommitRoundIOError
As part of CI tests in an unrelated PR, this test failed after detecting
a data race on the flag variable.  I couldn't reproduce the race again
on my own machine, but the code does seem like it can concurrently read
and write the bool flag, so it seems worth fixing to avoid spurious
data race warnings in the future.  This change replaces the flag with
an atomic.Bool instead.