catchup: make unsupported block handling more deterministic (#5673)
The catchup code had a race where, upon seeing an unsupported block
in the future, the catchup logic would cancel the current pipeline of
block fetches.  Depending on the timing, this might either make some
progress or not fetch any blocks at all.

The new logic is more deterministic: it tries to always fetch all blocks
up until the last supported block, and when the ledger reaches the last
supported block, the catchup service stops.