simplify single-leaf merklearray proofs
No need to explicitly specify tree depth for merklearray proofs of
a single leaf node, because the tree depth is exactly the number of
sibling hashes in the proof.

This allows simplifying the format for sending single-leaf merklearray
proofs, sending just the proofBytes []byte, and not having to send along
an extra treeDepth.

The current algod API for fetching a proof still sends the tree depth
along, for any clients that might care about it, but future APIs won't
need to pass treeDepth.