Put the three doc strings about each op together
This PR should not affect any run-time behavior, nor any generated
specs. It simply puts all of the documentation for an opcode together
in one place, so it's slightly easier to add opcodes and find all of
the documentation about an opcode in the source.

An eventual further change might be to get the documentaion into the
OpSpecs themselves, as there is still a little confusion there -
immediates are _named_ in the OpSpec, but documented in doc.go.