An opcode `heartbeat` that keeps an unlucky node online.
If, by chance, a node does not propose for a long time, they can issue
a transaction to explicitly indicate that they are still online, so
they won't be suspended. They must sign a block seed of recent block
using their particpation key and execute the `heartbeat` opcode on it.