Remove one allocate per hash by using generics.
The way we currently hash various objects is with:

```
// HashRep appends the correct hashid before the message to be hashed.
func HashRep(h Hashable) []byte {
	hashid, data := h.ToBeHashed()
	return append([]byte(hashid), data...)
}
```

This means that every callers generally have to allocate in order to
convert their argument, which might be a `BlockHeader`, for example,
into a Hashable. (This happens transparently, of course.)

However, by writing HashRep as:

```
func HashRep[H Hashable](h H) []byte {
	hashid, data := h.ToBeHashed()
	return append([]byte(hashid), data...)
}
```

We use generics to make a HashRep for each Hashable type. So a
`Hashable` need not be created to call it. Thus we we get one fewer
allocations most of the time.

For this PR, I did this by writing `HashRepFast` instead, so that I
could commit some benchmarks.  They show one for allocation.

I had to create several copies of existsing functions to make the
Benchmarks work.  In the real PR, I'll remove all that extra stuff.